satisfactory-planner-api/requirements.txt
fastapi==0.79.0
uvicorn==0.18.2
motor==3.0.0
python-decouple==3.6